Deformation motion of two-dimensional fish body
It is used for deformation movement of two-dimensional fish body to ensure that the length of the middle line of the fish body remains unchanged. The self-propulsion code for the S-shaped start of the fish, refer to the governing equations in the matlab file.  The code is based on the UDF modification of the following paper, welcome to read the citation.
Performance evaluation and optimization for twodimensional fish-like propulsion